requirements.txt
fastapi==0.133.1
openai==2.24.0
uvicorn==0.41.0
pydantic==2.12.5
python-dotenv==1.2.1
httpx==0.28.1
python-multipart==0.0.22
